change namespace again, extract extension, fix typo
